+++ templates/demo-store/app/routes/($lang).$shopid.orders.$token.authenticate.tsx
