+++ templates/demo-store/app/routes/($lang).account.activate.$id.$activationToken.tsx
