+++ templates/demo-store/app/routes/($lang).account.logout.ts
