+++ templates/demo-store/app/routes/($lang).account.orders.$id.tsx
