+++ templates/demo-store/app/routes/($lang).account.recover.tsx
