+++ templates/demo-store/app/routes/($lang).account.register.tsx
@@ -8,7 +8,7 @@ import {
 import {Form, useActionData} from '@remix-run/react';
 import {useState} from 'react';
 import {getInputStyleClasses} from '~/lib/utils';
-import {doLogin} from './login';
+import {doLogin} from './($lang).account.login';
 import type {CustomerCreatePayload} from '@shopify/hydrogen/storefront-api-types';
 import {Link} from '~/components';
