+++ templates/demo-store/app/routes/($lang).account.reset.$id.$resetToken.tsx
