+++ templates/demo-store/app/routes/($lang).api.products.tsx
