+++ templates/demo-store/app/routes/($lang).collections._index.tsx
