+++ templates/demo-store/app/routes/($lang).collections.all.tsx
