+++ templates/demo-store/app/routes/($lang).featured-products.tsx
