+++ templates/demo-store/app/routes/($lang).journal._index.tsx
