+++ templates/demo-store/app/routes/($lang).pages.$pageHandle.tsx
