+++ templates/demo-store/app/routes/($lang).policies.$policyHandle.tsx
