+++ templates/demo-store/app/routes/($lang).policies._index.tsx
