+++ templates/demo-store/app/routes/($lang).products.$productHandle.tsx
