+++ templates/demo-store/app/routes/($lang).products._index.tsx
