+++ templates/demo-store/app/routes/($lang).search.tsx
