+++ templates/demo-store/remix.config.js
@@ -21,5 +21,6 @@ module.exports = {
     unstable_tailwind: true,
     v2_errorBoundary: true,
     v2_meta: true,
+    v2_routeConvention: true,
   },
 };
